refactor - Mandando componentes reutilizados na aplicação para uma pasta commons.
